+++ extensions/Worker.Extensions.Rpc/release_notes.md
@@ -4,7 +4,6 @@
 - My change description (#PR/#issue)
 -->
 
-### Microsoft.Azure.Functions.Worker.Extensions.Rpc 1.0.0
+### Microsoft.Azure.Functions.Worker.Extensions.Rpc 1.0.1
 
-- Initial public release
-- Adds API for getting a `CallInvoker` pre-configured for communication with Functions host. 
+- Set max message send and receive length on gRPC `CallInvoker`.